♻️ refactor - 코드 컨벤션에 맞게 수정 및 옵티마이저
