+++ src/main/java/com/sparta/spartakanbanboard/domain/board/contoroller/BoardController.java
@@ -5,8 +5,8 @@ import com.sparta.spartakanbanboard.domain.board.dto.BoardResponseDto;
 import com.sparta.spartakanbanboard.domain.board.entity.Board;
 import com.sparta.spartakanbanboard.domain.board.service.BoardService;
 import com.sparta.spartakanbanboard.domain.user.entity.User;
-import com.sparta.spartakanbanboard.domain.user.service.global.dto.CommonResponseDto;
-import com.sparta.spartakanbanboard.domain.user.service.global.security.UserDetailsImpl;
+import com.sparta.spartakanbanboard.global.dto.CommonResponseDto;
+import com.sparta.spartakanbanboard.global.security.UserDetailsImpl;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Slice;
