+++ src/main/java/com/sparta/spartakanbanboard/domain/board/repository/BoardRepositoryQuery.java
@@ -2,7 +2,6 @@ package com.sparta.spartakanbanboard.domain.board.repository;
 
 import com.sparta.spartakanbanboard.domain.board.entity.Board;
 import com.sparta.spartakanbanboard.domain.user.entity.User;
-import com.sparta.spartakanbanboard.global.dto.PageDto;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
