+++ src/main/java/com/sparta/spartakanbanboard/domain/card/entity/Card.java
@@ -3,9 +3,8 @@ package com.sparta.spartakanbanboard.domain.card.entity;
 import com.sparta.spartakanbanboard.domain.card.dto.CreateCardRequestDto;
 import com.sparta.spartakanbanboard.domain.column.entity.KanbanColumn;
 import com.sparta.spartakanbanboard.domain.user.entity.User;
-import com.sparta.spartakanbanboard.domain.user.service.global.entity.TimeStamped;
+import com.sparta.spartakanbanboard.global.entity.TimeStamped;
 import jakarta.persistence.*;
-import jakarta.validation.constraints.NotBlank;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
