+++ src/main/java/com/sparta/spartakanbanboard/domain/card/service/CardService.java
@@ -1,7 +1,7 @@
 package com.sparta.spartakanbanboard.domain.card.service;
 
 import com.sparta.spartakanbanboard.domain.card.dto.CreateCardRequestDto;
-import com.sparta.spartakanbanboard.domain.user.service.global.dto.CommonResponseDto;
+import com.sparta.spartakanbanboard.global.dto.CommonResponseDto;
 
 public interface CardService {
