+++ src/main/java/com/sparta/spartakanbanboard/domain/user/entity/User.java
@@ -2,12 +2,13 @@ package com.sparta.spartakanbanboard.domain.user.entity;
 
 import com.sparta.spartakanbanboard.domain.card.entity.Card;
 import jakarta.persistence.*;
-import java.util.LinkedHashSet;
-import java.util.Set;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+
+import java.util.LinkedHashSet;
+import java.util.Set;
 
 @Getter
 @NoArgsConstructor
