+++ src/main/java/com/sparta/spartakanbanboard/domain/user/service/UserService.java
@@ -7,9 +7,9 @@ import com.sparta.spartakanbanboard.domain.user.entity.User;
 import com.sparta.spartakanbanboard.domain.user.entity.UserRole;
 import com.sparta.spartakanbanboard.domain.user.entity.UserStatus;
 import com.sparta.spartakanbanboard.domain.user.repository.UserRepository;
-import com.sparta.spartakanbanboard.domain.user.service.global.BusinessLogicException;
-import com.sparta.spartakanbanboard.domain.user.service.global.dto.CommonResponseDto;
-import com.sparta.spartakanbanboard.domain.user.service.global.jwt.JwtUtil;
+import com.sparta.spartakanbanboard.global.BusinessLogicException;
+import com.sparta.spartakanbanboard.global.dto.CommonResponseDto;
+import com.sparta.spartakanbanboard.global.jwt.JwtUtil;
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
