+++ src/main/java/com/sparta/spartakanbanboard/global/config/JpaConfig.java
@@ -1,4 +1,4 @@
-package com.sparta.spartakanbanboard.domain.user.service.global.config;
+package com.sparta.spartakanbanboard.global.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
