+++ src/main/java/com/sparta/spartakanbanboard/global/config/WebSecurityConfig.java
@@ -1,8 +1,8 @@
-package com.sparta.spartakanbanboard.domain.user.service.global.config;
+package com.sparta.spartakanbanboard.global.config;
 
-import com.sparta.spartakanbanboard.domain.user.service.global.security.UserDetailsServiceImpl;
-import com.sparta.spartakanbanboard.domain.user.service.global.jwt.JwtUtil;
-import com.sparta.spartakanbanboard.domain.user.service.global.security.JwtAuthorizationFilter;
+import com.sparta.spartakanbanboard.global.security.UserDetailsServiceImpl;
+import com.sparta.spartakanbanboard.global.jwt.JwtUtil;
+import com.sparta.spartakanbanboard.global.security.JwtAuthorizationFilter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
